Split contract.rs into logical sub-modules (queries, mutations)
The contract module has been restructured from a single file into a
directory-based module with separate files for:
- mod.rs: contract struct definition, helper functions, constants
- mutations.rs: state-changing functions (create, claim, complete)
- queries.rs: read-only functions (get_bounty, get_contributor, get_bounty_count)

This improves maintainability by making the separation of concerns
explicit at the file system level.

Closes #309
Closes #303
Closes #307
Closes #308
