+++ src/contract/queries.rs
@@ -1,0 +1,14 @@
+#[contractimpl]
+impl MergeMintContract {
+    pub fn get_bounty(env: Env, bounty_id: BytesN<32>) -> Option<Bounty> {
+        storage::get_bounty(&env, &bounty_id)
+    }
+
+    pub fn get_contributor(env: Env, address: Address) -> Option<Contributor> {
+        storage::get_contributor(&env, &address)
+    }
+
+    pub fn get_bounty_count(env: Env) -> u64 {
+        storage::get_bounty_count(&env)
+    }
+}
